+++ src/pages/Test.jsx
@@ -3,25 +3,29 @@ import { useQuery } from 'react-query';
 import { collection, getDocs } from 'firebase/firestore';
 import { db } from '../common/firebase';
 import { Spin } from 'antd';
+import usePosts from '../hooks/usePost';
 
 function Test() {
-  const { isLoading, isError, data, error } = useQuery(
-    'posts',
-    async () => {
-      const q = collection(db, 'post');
-      const querySnapshot = await getDocs(q);
-      const posts = querySnapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-      return posts;
-    },
-    {
-      cacheTime: 5 * 60 * 1000,
-      staleTime: 2 * 60 * 1000,
-      notifyOnChangeProps: ['data'], // 데이터가 바꾸지 않았을 때는 리렌더링 하지 않음
-    },
-  );
+  const { data, isLoading, isError, error } = usePosts();
+  console.log('🚀 ~ file: Test.jsx:10 ~ Test ~ data:', data);
+
+  // const { isLoading, isError, data, error } = useQuery(
+  //   'posts',
+  //   async () => {
+  //     const q = collection(db, 'post');
+  //     const querySnapshot = await getDocs(q);
+  //     const posts = querySnapshot.docs.map((doc) => ({
+  //       id: doc.id,
+  //       ...doc.data(),
+  //     }));
+  //     return posts;
+  //   },
+  //   {
+  //     cacheTime: 5 * 60 * 1000,
+  //     staleTime: 2 * 60 * 1000,
+  //     notifyOnChangeProps: ['data'], // 데이터가 바꾸지 않았을 때는 리렌더링 하지 않음
+  //   },
+  // );
 
   if (isLoading) {
     return <Spin />;
@@ -30,8 +34,6 @@ function Test() {
   if (isError) {
     return <div>Error: {error.message}</div>;
   }
-
-  console.log('🚀 ~ file: Test.jsx:24 ~ Test ~ data', data);
 
   return (
     <div>
