+++ src/pages/home/Home.jsx
@@ -17,8 +17,12 @@ import MocoChat from '../../components/mocoChat/MocoChatIcon';
 import { Modal } from 'antd';
 import AddInfoModal from '../../components/home/AddInfoModal';
 import { useNavigate } from 'react-router-dom';
+import usePosts from '../../hooks/usePost';
 
 const Home = () => {
+  const { data, isLoading, isError, error } = usePosts();
+  console.log('🚀 ~ file: Home.jsx:24 ~ Home ~ data:', data);
+
   const navigate = useNavigate();
   const currentUser = authService.currentUser;
   console.log('🚀 ~ file: Home.jsx:24 ~ Home ~ currentUser:', currentUser);
@@ -29,7 +33,7 @@ const Home = () => {
   const [isModalOpen, setIsModalOpen] = useState(false);
   //* 신규 유저 여부 상태
   const [isClosed, SetIsClosed] = useState(false);
-  const [postList, setPostList] = useState([]);
+  // const [postList, setPostList] = useState([]);
   const [userList, setUserList] = useState([]);
 
   // ! 추가 정보 등록 모달 핸들러
@@ -48,36 +52,49 @@ const Home = () => {
   const currentUserData = userList.filter(
     (item) => item.uid === currentUser?.uid,
   );
-  const recommendTechList = postList.filter(
-    (item) =>
-      !item.isDeleted &&
-      item.partyStack.includes(
-        currentUserData[0]?.moreInfo?.u_stack.toString(),
-      ),
-  );
-  const recommendTimeList = postList.filter(
-    (item) =>
-      !item.isDeleted &&
-      item.partyTime.includes(currentUserData[0]?.moreInfo?.u_time),
-  );
-  const recommendLocationList = postList.filter(
-    (item) =>
-      !item.isDeleted &&
-      item.partyLocation.includes(currentUserData[0]?.moreInfo?.u_location),
-  );
+  const recommendTechList = data
+    ? data.filter(
+        (item) =>
+          !item.isDeleted &&
+          item.partyStack.includes(
+            currentUserData[0]?.moreInfo?.u_stack.toString(),
+          ),
+      )
+    : [];
+  // const recommendTechList = data.filter(
+  //   (item) =>
+  //     !item.isDeleted &&
+  //     item.partyStack.includes(
+  //       currentUserData[0]?.moreInfo?.u_stack.toString(),
+  //     ),
+  // );
+  const recommendTimeList = data
+    ? data.filter(
+        (item) =>
+          !item.isDeleted &&
+          item.partyTime.includes(currentUserData[0]?.moreInfo?.u_time),
+      )
+    : [];
+  const recommendLocationList = data
+    ? data.filter(
+        (item) =>
+          !item.isDeleted &&
+          item.partyLocation.includes(currentUserData[0]?.moreInfo?.u_location),
+      )
+    : [];
 
-  useEffect(() => {
-    const postCollectionRef = collection(db, 'post');
-    const q = query(postCollectionRef, orderBy('createdAt', 'desc'));
-    const getPost = onSnapshot(q, (snapshot) => {
-      const postData = snapshot.docs.map((doc) => ({
-        id: doc.id,
-        ...doc.data(),
-      }));
-      setPostList(postData);
-    });
-    return getPost;
-  }, []);
+  // useEffect(() => {
+  //   const postCollectionRef = collection(db, 'post');
+  //   const q = query(postCollectionRef, orderBy('createdAt', 'desc'));
+  //   const getPost = onSnapshot(q, (snapshot) => {
+  //     const postData = snapshot.docs.map((doc) => ({
+  //       id: doc.id,
+  //       ...doc.data(),
+  //     }));
+  //     setPostList(postData);
+  //   });
+  //   return getPost;
+  // }, []);
 
   //postList -> 로그인 안 됐을 시 안보이게
   useEffect(() => {
@@ -104,7 +121,7 @@ const Home = () => {
         recommendTimeList={recommendTimeList}
         recommendLocationList={recommendLocationList}
       />
-      <HomeNewMeetingList postList={postList} />
+      <HomeNewMeetingList data={data} />
       <HomeAllBtn />
       {/* 신규 유저면 모달 오픈 */}
       <Modal open={isModalOpen} centered={true} closable={false} footer={false}>
